wireframe-backend/Wireframe-Generator/requirements.txt
# FastAPI and server
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20

# LangChain and LLM tools
langchain==0.3.25
langchain-core
langchain-google-genai==2.1.4
langchain-openai==0.3.17
langchain-anthropic==0.3.13
langgraph==0.4.5
langsmith==0.3.42

# Utilities
redis==6.1.0
httpx==0.28.1
tenacity==9.1.2

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
svgwrite==1.4.3
